[feat] 캐릭터 검색 컴포넌트의 최근검색 로직(로컬스토리지) 분리
