+++ src/app/profile/character/[characterName]/_component/CharacterDetailContent.module.scss
